Interactions overhaul

+++ rocon_hub/src/rocon_hub/redis_server.py
@@ -14,17 +14,21 @@ import shutil
 import subprocess
 import signal
 
-# Ros imports
+# Delete this once we upgrade (hopefully anything after precise)
+# Refer to https://github.com/robotics-in-concert/rocon_multimaster/issues/248
+import threading
+threading._DummyThread._Thread__stop = lambda x: 42
+
 import rospy
 import rospkg
 try:
-    import redis
+    import rocon_python_redis as redis
 except ImportError:
     # actually unused right now while we use redis as a ros package
     sys.exit("\n[ERROR] No python-redis found - 'rosdep install rocon_hub'\n")
+import rocon_semantic_version as semantic_version
 
-# Local imports
-import utils
+from . import utils
 
 ##############################################################################
 # Redis Server
@@ -40,29 +44,59 @@ class RedisServer:
         if os.path.isdir(self._home_dir):
             shutil.rmtree(self._home_dir)
         self._files = {}
-        self._files['redis_conf'] = os.path.join(self._home_dir, 'redis.conf')
-        self._files['redis_conf_local'] = os.path.join(self._home_dir, 'redis.conf.local')
+        self._version_extension = self._introspect_redis_server_version()
+        self._files['redis_conf'] = os.path.join(self._home_dir, 'redis-%s.conf' % self._version_extension)
+        self._files['redis_conf_local'] = os.path.join(self._home_dir, 'redis-%s.conf.local' % self._version_extension)
         self._files['redis_server_log'] = os.path.join(self._home_dir, 'redis-server.log')
         self._server = None
-
         self._setup()
+
+    def _introspect_redis_server_version(self):
+        '''
+          Sniff the version in major.minor format for decision making elsewhere (patch we disregard since our
+          decisions don't depend on it).
+
+          @return version extension in 'major.minor' format.
+          @rtype str
+        '''
+        process = subprocess.Popen(["redis-server", "--version"], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        output, unused_error = process.communicate()
+        try:
+            version_string = re.search('v=([0-9.]+)', output).group(1)  # 2.6+ pattern
+        except AttributeError:
+            version_string = re.search('version ([0-9.]+)', output).group(1)  # 2.2 pattern
+        version = semantic_version.Version(version_string)
+        rospy.loginfo("Hub : version %s" % (version_string))
+        return str(version.major) + "." + str(version.minor)
 
     def _setup(self):
         '''
           Clear and configure redis conf, log files in the ros home
           directories under a subdirectory styled by the name of this hub.
+
+          Also check that we have support for the redis server - i.e. check if we
+          have a .conf file for that version and exit this script if not found.
         '''
         if os.path.isdir(self._home_dir):
             shutil.rmtree(self._home_dir)
         os.makedirs(self._home_dir)
         rospack = rospkg.RosPack()
-        redis_conf_template = utils.read_template(os.path.join(rospack.get_path('rocon_hub'), 'redis', 'redis.conf'))
-        redis_conf_template = instantiate_redis_conf_template(redis_conf_template, self._files['redis_conf_local'])
-        redis_local_template = utils.read_template(os.path.join(rospack.get_path('rocon_hub'), 'redis', 'redis.conf.local'))
+        package_redis_conf_file = os.path.join(rospack.get_path('rocon_hub'), 'redis', 'redis-%s.conf' % self._version_extension)
+        package_redis_conf_local_file = os.path.join(rospack.get_path('rocon_hub'), 'redis', 'redis-%s.conf.local' % self._version_extension)
+
+        # Checks
+        if not os.path.isfile(package_redis_conf_file):
+            utils.logfatal("Hub : the version of the redis server you have installed is not supported by rocon.")
+            sys.exit(utils.logfatal("Hub : please submit a ticket at https://github.com/robotics-in-concert/rocon_multimaster"))
+
+        redis_conf_template = utils.read_template(package_redis_conf_file)
+        redis_conf_template = instantiate_redis_conf_template(redis_conf_template, self._files['redis_conf_local'])  # drop the local file path to use into the settings
+        redis_local_template = utils.read_template(package_redis_conf_local_file)
         redis_local_template = instantiate_local_conf_template(redis_local_template,
                                                                self._parameters['port'],
                                                                self._parameters['max_memory'],
-                                                               self._files['redis_server_log'])
+                                                               self._files['redis_server_log'],
+                                                               self._home_dir)
         try:
             f = open(self._files['redis_conf'], 'w')
             f.write(redis_conf_template.encode('utf-8'))
@@ -152,7 +186,7 @@ def instantiate_redis_conf_template(template, local_conf_filename):
     return template % locals()
 
 
-def instantiate_local_conf_template(template, port, max_memory, logfile):
+def instantiate_local_conf_template(template, port, max_memory, logfile, working_dir):
     '''
       Variable substitution in a template file.
 
@@ -162,6 +196,10 @@ def instantiate_local_conf_template(template, port, max_memory, logfile):
       @type string
       @param max_memory: how much memory to allocate to the redis server in bytes
       @type string (e.g. 10mb)
+      @param logfile
+      @type string
+      @param working_dir : filesystem which redis uses to dump (can we turn this off in 2.6?)
+      @type string
     '''
     return template % locals()
